requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
faiss-cpu
google-generativeai
chromadb
langchain_google_genai